Let BSPWM revomal_adjustment work.
The removal_adjustment setting of BSPWM, when set, rearranges the
brother node of the node that has been closed or moved. It doesn't take
place when the node is hidden, though. Thus, before we hide the node, we
move it to a temporary workspace, so that removal_adjustment takes
place.